gst-omx: Optimize the startup speed of some streams
When the reconfiguration happened, need to return GST_OMX_ACQUIRE_BUFFER_RECONFIGURE
as soon as possible, especially when the filled length of pending buffer is
zero. Because the OMX il will return the output buffer in the state of reconfiguration.
It makes the output buffer been pushed back and forth between the gstomx and OMX il,
which makes the reconfiguration too later.

original patch: 3533881

Change-Id: I84c4cf2b80145ba43a37f633b8b4778f14879f18
Signed-off-by: Lily Li <quic_lali@quicinc.com>