gst-omx: dec: reallocate input buffer for external input mode
Previously, we avoided reallocating the input buffer during output
port reconfiguration when the buffer was externally allocated, due
to a bug in the OMX module that caused the input buffer fd to be
closed unexpectedly in omx module. This bug has been fixed recently.
Additionally, there was a failure when updating the input port
definition. The root cause was the failing of requesting V4L2 input
buffers while the input port was still streaming. Reallocating the
input buffer resolves this issue, as the STREAMOFF ioctl command is
triggered during the reallocation process, which stopped the stream.

So, it needs to reallocate omxil internal input buffer for external
input mode when reconfiguring output port.

Change-Id: I2da3bd4d993ee60019074b8e67d16ea4104d84a8
Signed-off-by: chlei <chlei@qti.qualcomm.com>